Module 3: Lab Activity
This activity circles back to our first module, where we had to decipher the errors of the code and figure out what the errors were in the code. Once we found all possible errors, we needed to test run the code to see if it functioned properly, resulting in trial and error. We needed to write at least a paragraph-long explanation about the mistakes we saw in our assignment. We then learned why documenting the code and its errors was beneficial during coding and why we needed to perform this process before running the program, during coding, and after launching it.